Refactor code to be shared
